requirements.txt
python-telegram-bot
python-dotenv
openai
fuzzywuzzy
python-Levenshtein
requests
beautifulsoup4